\name{dynpred-package}
\alias{dynpred-package}
\alias{dynpred}
\docType{package}
\title{
The companion package of the book "Dynamic Prediction in Survival Analysis"
}
\description{
The companion package of the book "Dynamic Prediction in Survival Analysis".}
\details{
\tabular{ll}{
Package: \tab dynpred\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2011-11-01\cr
License: \tab GPL 2.0\cr
}
An overview of how to use the package, including the most important
functions.
}
\author{
Hein Putter
Maintainer: Hein Putter <H.Putter@lumc.nl>
}
\references{
van Houwelingen HC, Putter H (2011). Dynamic Prediction in Clinical Survival Analysis.
Chapman & Hall.
}
\keyword{ package }
