% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prederror.R
\name{pe}
\alias{pe}
\alias{pecox}
\title{Calculate prediction error curve}
\usage{
pe(time, status, tsurv, survmat, tcens, censmat, FUN = c("KL", "Brier"), tout)

pecox(formula, censformula, data, censdata, FUN = c("KL", "Brier"), tout,
  CV = FALSE, progress = FALSE)
}
\arguments{
\item{time}{Vector of time points in data}

\item{status}{Vector of event indicators in data}

\item{tsurv}{Vector of time points corresponding to the estimated survival
probabilities in \code{survmat}}

\item{survmat}{Matrix of estimated survival probabilities; dimension should
be length of tsurv x length of time}

\item{tcens}{Vector of time points corresponding to the estimated censoring
probabilities in \code{censmat}}

\item{censmat}{Matrix of estimated censoring probabilities; dimension should
be length of tcens x length of time}

\item{FUN}{The error function, either \code{"KL"} (default) for
Kullback-Leibler or \code{"Brier"} for Brier score}

\item{tout}{Vector of time points at which to evaluate prediction error. If
missing, prediction error will be evaluated at all time points where the
estimate will change value}

\item{formula}{Formula for prediction model to be used as in
\code{\link[survival:coxph]{coxph}}}

\item{censformula}{Formula for censoring model, also to be used as in
\code{\link[survival:coxph]{coxph}}}

\item{data}{Data set in which to interpret \code{formula}}

\item{censdata}{Data set in which to interpret \code{censformula}}

\item{CV}{Boolean (default=\code{FALSE}); if \code{TRUE}, (leave-one-out)
cross-validation is used for the survival probabilities}

\item{progress}{Boolean (default=\code{FALSE}); if \code{TRUE}, progress is
printed on screen}
}
\value{
A data frame with columns \item{time}{Event time points}
\item{Err}{Prediction error of model specified by \code{formula} at these
time points}
}
\description{
Calculate prediction error curve.
}
\details{
The \code{censformula} is used to calculate inverse probability of censoring
weights (IPCW).
}
\examples{
data(ova)
# Example on a subset, because the effect of CV is clearer
ova2 <- ova[1:100,]
pecox(Surv(tyears, d) ~ Karn + Broders + FIGO + Ascites + Diam, Surv(tyears, 1-d) ~ 1,
  data = ova2, FUN="Brier", tout=seq(0,6,by=0.5))
pecox(Surv(tyears, d) ~ Karn + Broders + FIGO + Ascites + Diam, Surv(tyears, 1-d) ~ 1,
  data = ova2, FUN="Brier", tout=seq(0,6,by=0.5), CV=TRUE, progress=TRUE)

\donttest{
pecox(Surv(tyears, d) ~ Karn + Broders + FIGO + Ascites + Diam, Surv(tyears, 1-d) ~ 1,
  data = ova, FUN="Brier", tout=seq(0,6,by=0.5))
pecox(Surv(tyears, d) ~ Karn + Broders + FIGO + Ascites + Diam, Surv(tyears, 1-d) ~ 1,
  data = ova, FUN="Brier", tout=seq(0,6,by=0.5), CV=TRUE, progress=TRUE)
}
}
\author{
Hein Putter \email{H.Putter@lumc.nl}
}
\references{
Graf E, Schmoor C, Sauerbrei W & Schumacher M (1999), Assessment
and comparison of prognostic classification schemes for survival data,
Statistics in Medicine 18, 2529-2545.

Gerds & Schumacher (2006), Consistent estimation of the expected Brier score
in general survival models with right-censored event times, Biometrical
Journal 48, 1029-1040.

van Houwelingen HC, Putter H (2012). Dynamic Prediction in Clinical Survival
Analysis. Chapman & Hall.
}
\keyword{univar}

