% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynpred-package.R
\docType{package}
\name{dynpred-package}
\alias{dynpred}
\alias{dynpred-package}
\title{The companion package of the book "Dynamic Prediction in Survival Analysis"}
\description{
The companion package of the book "Dynamic Prediction in Survival Analysis".
}
\details{
\tabular{ll}{ Package: \tab dynpred\cr Type: \tab Package\cr Version: \tab
0.1.2\cr Date: \tab 2014-11-10\cr License: \tab GPL (>= 2)\cr } An overview
of how to use the package, including the most important functions.
}
\author{
Hein Putter Maintainer: Hein Putter <H.Putter@lumc.nl>
}
\references{
van Houwelingen HC, Putter H (2012). Dynamic Prediction in
Clinical Survival Analysis. Chapman & Hall.
}
\keyword{package}

